requirements.txt
discord.py==2.4.0
toml==0.10.2
pynacl==1.5.0
